rename: 훅 파일 이름 표기법 변경
